Seperate resolver files
