+++ src/db/resolvers/comment_resolver.js
@@ -1,0 +1,24 @@
+const commentTypeDefs = ``;
+
+const commentResolver = {
+  Query: {
+    createNewComment(_, args) {
+      return "success";
+    },
+    updateComment(_, args) {
+      return "success";
+    },
+    deleteComment(_, args) {
+      return "success";
+    },
+    voteComment(_, args) {
+      return "success"
+    }
+  }
+};
+
+
+module.exports = {
+  commentTypeDefs,
+  commentResolver,
+}
