+++ src/db/resolvers/user_resolver.js
@@ -1,0 +1,15 @@
+const userTypeDefs = ``;
+
+const userResolver = {
+  Query: {
+    userLogin(_, args) {
+      return "success";
+    },
+  }
+};
+
+
+module.exports = {
+  userTypeDefs,
+  userResolver,
+}
